Avanzamento

+++ jar/src/main/java/lispa/schedulers/dao/sgr/sire/history/SireHistoryWorkitemDAO.java
@@ -613,7 +613,7 @@ public class SireHistoryWorkitemDAO {
 			throws Exception {
 		ConnectionManager cm = null;
 		Connection connection = null;
-		List<Long> listPk = new ArrayList<>();
+		List<Integer> listPk = new ArrayList<>();
 		try {
 			cm = ConnectionManager.getInstance();
 			connection = cm.getConnectionOracle();
@@ -633,8 +633,7 @@ public class SireHistoryWorkitemDAO {
 
 				try (ResultSet rs = ps2.executeQuery();) {
 					while (rs.next()) {
-						
-						listPk.add(rs.getLong(1));
+						listPk.add(rs.getInt(1));
 					}
 					int numberInParameters = listPk.size();
 					String parameters = "";
@@ -653,7 +652,7 @@ public class SireHistoryWorkitemDAO {
 						// Array array=connection.createArrayOf("NUMBER",
 						// listPk.toArray());
 						for (int i = 0; i < listPk.size(); i++) {
-							ps.setLong(i + 2, listPk.get(i));
+							ps.setInt(i + 2, listPk.get(i));
 						}
 						ps.setTimestamp(listPk.size() + 2, DataEsecuzione
 								.getInstance().getDataEsecuzione());
